// typescript/trading-economics-dashboard/types/country-data.ts
export interface CountryData {
  Country: string;
  Category: string;
  Title: string;
  LatestValueDate: string;
  LatestValue: number;
  Source: string;
  SourceURL: string;
  Unit: string;
  URL: string;
  CategoryGroup: string;
  Adjustment: string;
  Frequency: string;
  HistoricalDataSymbol: string;
  CreateDate: string;
  FirstValueDate: string;
  PreviousValue: number;
  PreviousValueDate: string;
}